增加了一个multicore参数，用于在调试时关闭平行计算

+++ src/RcppExports.cpp
@@ -6,12 +6,12 @@
 using namespace Rcpp;
 
 // nearest_neighbors
-List nearest_neighbors(NumericVector xy, int rmax);
+List nearest_neighbors(NumericVector xy, double rmax);
 RcppExport SEXP sce_nearest_neighbors(SEXP xySEXP, SEXP rmaxSEXP) {
 BEGIN_RCPP
     Rcpp::RNGScope __rngScope;
     NumericVector xy = Rcpp::as<NumericVector >(xySEXP);
-    int rmax = Rcpp::as<int >(rmaxSEXP);
+    double rmax = Rcpp::as<double >(rmaxSEXP);
     List __result = nearest_neighbors(xy, rmax);
     return Rcpp::wrap(__result);
 END_RCPP
